Added Seeds files, with geolocation, created file geolocation_data_exploring, to future implemente visualization with coordinates; Updated Staging tables, tests, sources and schemas. Updated DBT staging README with what was done in the Staging Layer. Added staging markdown docs.

+++ RegTech/dbt_transform/models/staging/stg__dim_pix.sql
@@ -8,7 +8,7 @@ WITH dim_pix_source AS (
 SELECT
     "Nome Reduzido" AS nome_reduzido,
     "ISPB" AS ispb,
-    REPLACE(REPLACE(SUBSTRING("CNPJ" FROM 1 FOR POSITION('/' IN "CNPJ") - 1), '.', ''), '-', '') AS cnpj,
+    REPLACE(REPLACE(SUBSTRING("CNPJ" FROM 1 FOR POSITION('/' IN "CNPJ") - 1), '.', ''), '-', '')::INTEGER AS cnpj,
     "Tipo de Instituição" AS tipo_de_instituicao,
     "Autorizada pelo BCB" AS autorizada_bcb,
     "Tipo de Participação no SPI" AS tipo_participacao_spi,
